[Feature] add trait for entry and refactor project arrangement
