+++ lib/storage/src/entry/error.rs
@@ -1,5 +1,8 @@
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
+    #[error("invalid magic")]
+    InvalidMagic,
+
     #[error("prost encode")]
     ProstEncode(#[from] prost::EncodeError),
