+++ lib/storage/src/journal/error.rs
@@ -1,0 +1,2 @@
+#[derive(thiserror::Error, Debug)]
+pub enum Error {}
