+++ lib/storage/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod entry;
+pub mod journal;
